requirements.txt
pillow
pillow_heif
ffmpeg